+++ schema-kenerator-jsonschema/src/main/kotlin/io/github/smiley4/schemakenerator/jsonschema/JsonSchemaCompileReferenceStep.kt
@@ -145,6 +145,10 @@ internal class JsonSchemaCompileReferenceStep(
                     this.putAll(refObj.properties)
                     this.remove("\$ref")
                 })
+                // cleanup min/exclusiveMin, max/exclusiveMax
+                if (it.properties.contains("exclusiveMinimum")) { it.properties.remove("minimum") }
+                if (it.properties.contains("exclusiveMaximum")) { it.properties.remove("maximum") }
+                // cleanup nullable
                 if (it.properties.contains("_nullable") && it.getBool("_nullable") && explicitNullTypes) {
                     setNullable(it)
                     it.properties.remove("_nullable")