+++ schema-kenerator-jsonschema/src/main/kotlin/io/github/smiley4/schemakenerator/jsonschema/JsonSchemaCoreAnnotationDefaultStep.kt
@@ -10,6 +10,9 @@ import io.github.smiley4.schemakenerator.jsonschema.data.JsonSchemaData
 import io.github.smiley4.schemakenerator.jsonschema.jsonDsl.JsonObject
 import io.github.smiley4.schemakenerator.jsonschema.jsonDsl.JsonTextValue
 
+/**
+ * Handles the core [Default] annotation
+ */
 internal class JsonSchemaCoreAnnotationDefaultStep {
 
     fun process(input: IntermediateJsonSchemaData): IntermediateJsonSchemaData {