+++ schema-kenerator-reflection/src/main/kotlin/io/github/smiley4/schemakenerator/reflection/analyzer/SimpleTypeAnalyzerModule.kt
@@ -33,7 +33,8 @@ class SimpleTypeAnalyzerModule(
      */
     private fun KClass<*>.toTypeName() = TypeName(
         full = this.qualifiedName ?: this.java.name,
-        short = this.simpleName ?: this.java.name
+        short = this.simpleName ?: this.java.name,
+        packageName = this.java.packageName
     )
 
 }