+++ schema-kenerator-serialization/src/main/kotlin/io/github/smiley4/schemakenerator/serialization/analyzer/serialDescriptorUtils.kt
@@ -3,6 +3,7 @@ package io.github.smiley4.schemakenerator.serialization.analyzer
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.SerialDescriptor
+import kotlinx.serialization.descriptors.nonNullOriginal
 import kotlinx.serialization.serializerOrNull
 import kotlin.reflect.KClass
 import kotlin.reflect.KType
@@ -11,7 +12,7 @@ import kotlin.reflect.KType
  * @return the name for this serial descriptor without modifiers (e.g. nullable)
  */
 @OptIn(ExperimentalSerializationApi::class)
-fun SerialDescriptor.fullName() = this.serialName.replace("?", "")
+fun SerialDescriptor.fullName() = this.nonNullOriginal.serialName
 
 /**
  * Get the serializer (or null) for the given type