+++ tests/src/test/kotlin/io/github/smiley4/schemakenerator/test/cases/CoreAnnotationTestCases.kt
@@ -4,11 +4,19 @@ import io.github.smiley4.schemakenerator.core.annotations.Default
 import io.github.smiley4.schemakenerator.core.annotations.Deprecated
 import io.github.smiley4.schemakenerator.core.annotations.Description
 import io.github.smiley4.schemakenerator.core.annotations.Example
+import io.github.smiley4.schemakenerator.core.annotations.ExclusiveMax
+import io.github.smiley4.schemakenerator.core.annotations.ExclusiveMin
 import io.github.smiley4.schemakenerator.core.annotations.Format
+import io.github.smiley4.schemakenerator.core.annotations.Max
+import io.github.smiley4.schemakenerator.core.annotations.MaxLength
+import io.github.smiley4.schemakenerator.core.annotations.Min
+import io.github.smiley4.schemakenerator.core.annotations.MinLength
+import io.github.smiley4.schemakenerator.core.annotations.Pattern
 import io.github.smiley4.schemakenerator.core.annotations.Title
 import io.github.smiley4.schemakenerator.jsonschema.JsonSchemaSteps.handleCoreAnnotations
 import io.github.smiley4.schemakenerator.swagger.SwaggerSteps.handleCoreAnnotations
 import kotlinx.serialization.Serializable
+import org.checkerframework.common.value.qual.MinLen
 import kotlin.reflect.typeOf
 
 object CoreAnnotationTestCases {
@@ -29,7 +37,10 @@ object CoreAnnotationTestCases {
                   "title": "Annotated Class",
                   "required": [
                       "stringValue",
-                      "intValue"
+                      "intValue",
+                      "constrainedNumber",
+                       "constrainedString",
+                        "constrainedList"
                   ],
                   "type": "object",
                   "properties": {
@@ -46,6 +57,26 @@ object CoreAnnotationTestCases {
                           "description": "Int field description",
                           "default": "1111",
                           "example": "2222"
+                      },
+                      "constrainedNumber" : {
+                        "type" : "integer",
+                        "format" : "int32",
+                        "exclusiveMaximum" : 7,
+                        "minimum" : 2
+                      },
+                      "constrainedString" : {
+                        "type" : "string",
+                        "maxLength" : 5,
+                        "minLength" : 2,
+                        "pattern" : "[0-9]*"
+                      },
+                      "constrainedList" : {
+                        "type" : "array",
+                        "items" : {
+                          "type" : "boolean"
+                        },
+                        "maxItems" : 8,
+                        "minItems" : 4
                       }
                   },
                   "description": "some description",
@@ -61,7 +92,7 @@ object CoreAnnotationTestCases {
         expectedJson = """
             {
               "type": "object",
-              "required": ["stringValue", "intValue"],
+              "required": ["stringValue", "intValue", "constrainedNumber", "constrainedString", "constrainedList"],
               "properties": {
                   "stringValue": {
                       "type": "string",
@@ -82,6 +113,25 @@ object CoreAnnotationTestCases {
                           "2222"
                       ],
                       "format": "int32"
+                  },
+                  "constrainedNumber": {
+                     "type": "integer",
+                     "minimum": 2,
+                     "exclusiveMaximum": 7
+                  },
+                  "constrainedString": {
+                     "type": "string",
+                     "minLength": 2,
+                     "maxLength": 5,
+                     "pattern": "[0-9]*"
+                  },
+                  "constrainedList": {
+                     "type": "array",
+                     "items": {
+                        "type": "boolean"
+                     },
+                     "minItems": 4,
+                     "maxItems": 8
                   }
               },
               "title": "Annotated Class",
@@ -187,6 +237,20 @@ object CoreAnnotationTestCases {
         @Example("2222")
         @Format("int32")
         val intValue: Int,
+
+        @Min(2)
+        @Max(5)
+        @ExclusiveMax(7)
+        val constrainedNumber: Int,
+
+        @MinLength(2)
+        @MaxLength(5)
+        @Pattern("[0-9]*")
+        val constrainedString: String,
+
+        @MinLength(4)
+        @MaxLength(8)
+        val constrainedList: List<Boolean>,
     )